+++ packages/browser/src/Session.ts
@@ -61,6 +61,7 @@ export class Session extends EventEmitter {
   public readonly info: ISessionInfo;
 
   private clientAuthentication: ClientAuthentication;
+  private tokenRequestInProgress = false;
 
   /**
    * Session object constructor. Typically called as follows:
@@ -150,6 +151,13 @@ export class Session extends EventEmitter {
   handleIncomingRedirect = async (
     url: string
   ): Promise<ISessionInfo | undefined> => {
+    if (this.info.isLoggedIn) {
+      return this.info;
+    }
+    if (this.tokenRequestInProgress) {
+      return undefined;
+    }
+    this.tokenRequestInProgress = true;
     const sessionInfo = await this.clientAuthentication.handleIncomingRedirect(
       url
     );
@@ -163,6 +171,7 @@ export class Session extends EventEmitter {
       this.info.webId = sessionInfo.webId;
       this.info.sessionId = sessionInfo.sessionId;
     }
+    this.tokenRequestInProgress = false;
     return sessionInfo;
   };